\name{statystykaip}
\alias{statystykaip}
\title{Funkcje dostepne w bibliotece statystykaip}
\description{Funkcje}

\usage{
KWANTYLE:
kwantyl_rzedu(proba, rzad)
rozstep_kwartylowy(proba)

DLA ZDARZEN JEDNAKOWO PODOBNYCH (ELM)
srednia_arytmetyczna(proba)
odchylenie_standardowe(proba)
moment_rzedu(proba, rzad)
wariancja(proba)
wariancja_probkowa(proba)

DLA ROZKLADU Xi Pi:
wariancja_rozkladu(proba, rozklad)
wartosc_oczekiwana_rozkladu(proba, rozklad)

ASYMETRIA:
wsp_asymetrii_As(proba)
wsp_asymetrii_Aq(proba)
asymetria(proba)
wsp_zmiennosci(proba)
}

\author{Bartosz Zukowski}