\name{wartosc_oczekiwana_rozkladu}
\alias{wartosc_oczekiwana_rozkladu}
\title{
  Oblicza wartość oczekiwaną rozkładu losowego dla danej próby i jej rozkładu.
}
\usage{
  wartosc_oczekiwana_rozkladu(proba, rozklad, na.rm = TRUE)
}
\arguments{
  \item{proba}{
    Wektor numeryczny reprezentujący próbę.
  }
  \item{rozklad}{
    Wektor numeryczny reprezentujący rozkład prawdopodobieństwa odpowiadający próbie.
  }
  \item{na.rm}{
    Logiczna flaga określająca, czy należy usunąć brakujące wartości z wektorów wejściowych.
    Domyślnie \code{TRUE}.
  }
}
\description{
  Funkcja oblicza wartość oczekiwaną rozkładu losowego dla danej próby i jej rozkładu.
}
\examples{

}
